% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_demultiplex_fastq}
\alias{EMOTE_demultiplex_fastq}
\title{Demultiplex a EMOTE FASTQ file according to barcode sequence of the reads.}
\usage{
EMOTE_demultiplex_fastq(fq.file, out.dir = paste0(fq.file, ".demux"),
  force = FALSE, yieldSize = 1e+06, invalid.reads = TRUE, ...)
}
\arguments{
\item{fq.file}{a character with the path of the FASTQ file (eventually
gzipped) containing single-end EMOTE reads}

\item{out.dir}{a character with the path of the output directory where the
demultiplexed FASTQ files will be stored}

\item{force}{when TRUE recursively remove the content of out.dir before
running instead of stopping the function with an error message}

\item{yieldSize}{number of read to process per iteration}

\item{invalid.reads}{logical saying whether invalid reads should be outputed
in the file INVALID.fastq.gz}

\item{...}{additional parameters are passed to EMOTE_parse_reads}
}
\value{
a data.frame with demultiplexing statistics
}
\description{
Parse the reads from the input FASTQ file and store valid mapping sequences
in separate FASTQ files in the given output directory. The name of the file
in the ouput directory depends on the barcode sequence. At the end of the
process demultiplexing statistics are reported in the file
demultiplex_report.txt.
}
\details{
When the file demultiplex_report.txt already exists in the output directory
and force=FALSE, no demultiplexing is performed, but the report is load and
the function return with a warning message. When force=TRUE, the output
directory is recursively deleted and the demultiplex process if forced to
run.

By default the function processes the input fastq by chunks of 1e6 reads.
The number of read processed by the function can be tuned with the parameter
yieldSize to control memory usage.
}
