% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_map}
\alias{EMOTE_map}
\title{Map reads to a genome with Rbowtie}
\usage{
EMOTE_map(bowtie_index, fq.file, bam.file = sub("(.fastq|.fq)(.gz)?$",
  ".bam", fq.file), intermediate.fastq.file = tempfile(fileext =
  ".fastq"), force = FALSE, threads = 3, yieldSize = 1e+06,
  bowtie_v = 1, ...)
}
\arguments{
\item{bowtie_index}{a length-one character whith the name of the bowtie index
onto which the reads have to be mapped}

\item{fq.file}{name of the EMOTE FASTQ file to map (usualy a demultiplexed FASTQ)}

\item{bam.file}{name of the generated BAM file}

\item{intermediate.fastq.file}{name of the intermediate FASTQ file that is generated}

\item{force}{when TRUE the BAM file is generated wven if it already exists}

\item{threads}{number of thread to be used by Rbowtie during the mapping process}

\item{yieldSize}{the number of read processed per iteration when parsing the input FASTQ file to generate the intermediate file}

\item{bowtie_v}{the v parameter to bowtie (no more than V mismatches in the alignment)}

\item{...}{additional paramaters passed to EMOTE_parse_reads()}
}
\value{
the name of the generated BAM file
}
\description{
Parse the EMOTE reads in the given FASTQ file to extract the mapping sequence
and the UMI. Create an intermediate FASTQ file with this mapping sequence and
the UMI sequence added as a prefix of the reads headers. Map the intermediate
FASTQ file with Rbowtie and generate a BAM.
}
\details{
When the BAM file to generate already exists, the function simply return its name
}
\seealso{
bowtie_build
}
