% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_parse_reads}
\alias{EMOTE_parse_reads}
\title{Parse reads generated with EMOTE protocol}
\usage{
EMOTE_parse_reads(sr, max.recognition.mismatch = 1L,
  valid.barcodes = DNAStringSet(c("TACA", "GTAT", "CGTC", "AAGT", "ACAC",
  "GGTA", "GCCT", "TCGG", "CAAG", "TTGA", "GCTG", "CCGA", "CTCG", "AGGA",
  "ATTG", "GACG", "TGTT")), first.nuc.len = 1L,
  recognition.sequence = "CGGCACCAACCGAGG", umi.length = 7L,
  control.sequence = "CGC")
}
\arguments{
\item{sr}{a ShortRead object to parse}

\item{max.recognition.mismatch}{number of mismatch allowed in the recognition
sequence}

\item{valid.barcodes}{a character vector listing valid barcodes. If NULL, all
barcodes are considered valid.}

\item{first.nuc.len}{length of the first nucleotide sequence before EMOTE
recognition sequence start}

\item{recognition.sequence}{a character with the recognition sequence to
detect in the reads}

\item{umi.length}{length of the UMI sequence}

\item{control.sequence}{a character specifying the control sequence to look
for in the read}
}
\value{
A DataFrame where the columns contains the elements parsed in the
        input reads or a logical checking for the validity of the elements.
        The fields are:
 \itemize{
       \item first_nt: the first nucleotide of the read;
       \item barcode: the 4 nucleotides at position 2-5 in the read, which is
             expected to be the the barcode sequence;
       \item recognition: the 15 nucleotides at position 6-20 in the read,
             which is expected to be the EMOTE recognition sequence equal to
             CGGCACCAACCGAGG;
       \item umi: the 7 nucleotides at position 21-27 in the read, which is
             expected to be an universal molecular identifier (UMI);
       \item control: the 3 nucleotides at position 28-30 in the read, which
             is expected to be equal to the EMOTE control sequence ACG;
       \item read: the end of the read from position 31 which is the read
             sequence expected to map on the genome
       \item is_valid_recognition: a logical which is TRUE when the
             reconigiton sequence match with the sequence CGGCACCAACCGAGG
             with max.mismatch error tolerance;
       \item is_valid_control: a logical which is TRUE when the control
             sequence perfectly match with the sequence CGC;
       \item is_valid_umi: a logical which is TRUE when the UMI sequence only
             contains A,C or G nucleotides;
       \item is_valid_barcode: a logical which is TRUE when the barcode
             perfectly match with one of the given valid.barcodes, or when
             valid.barcodes is NULL;
       \item is_valid_EMOTE_read: a logical which is TRUE when
             is_valid_recognition, is_valid_control, is_valid_umi, and
             is_valid_barcode are all TRUE.
}
}
\description{
The function takes the given ShortRead object, split the reads into elements
expected to by found in EMOTE reads, and test for their validity.
}
