% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_quantify}
\alias{EMOTE_quantify}
\title{Quantify number of read starting at each position of the genome}
\usage{
EMOTE_quantify(bam.file, quantif.file = paste0(bam.file, ".quantif"),
  remove.umi.duplicate = TRUE, mode = c("unambiguous", "ambiguous", "all"),
  yieldSize = NA_integer_, force = FALSE)
}
\arguments{
\item{bam.file}{the input bam files to process}

\item{quantif.file}{name of the output directory that will contain the quantification results}

\item{remove.umi.duplicate}{a logical telling whether the umi correction must be done}

\item{mode}{the quantification mode to use: either all reads, either only ambiguous reads or only unambiguous reads.
Reads with a mapping quality of 0 are considered ambiguously mapped.}

\item{yieldSize}{yieldSize value passed to BamFile to process input file by chunks when limited amount of memory is available}

\item{force}{logical, when TRUE erase existing quantification and regenerate a new one with new parameters}
}
\value{
name of the generated quantification directory
}
\description{
Quantify number of read starting at each position of the genome with optional
UMI correction, and store the result in an output directory
}
