% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_read_quantif}
\alias{EMOTE_read_quantif}
\title{Load one or multiple EMOTE quantification files}
\usage{
EMOTE_read_quantif(quantif.files)
}
\arguments{
\item{quantif.files}{a character vector listing the quantification directories to load}
}
\value{
a SummarizedExperiment object containing the quantification values for all samples
        at each non-empty genomic location
}
\description{
Load one or multiple EMOTE quantification files
}
