% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMOTEconv.R
\name{EMOTE_readConvOutput}
\alias{EMOTE_readConvOutput}
\title{Read EMOTE data from a ZIP archive containing multiple EmoteConv output directories}
\usage{
EMOTE_readConvOutput(zip.file, ambig = FALSE)
}
\arguments{
\item{zip.file}{filename of the ZIP archive to load. The ZIP archive may contain multiple directories generated by EmoteConv.}

\item{ambig}{a logical specifying weither to load ambiguous or unambiguously counts}
}
\value{
a SummarizedExperiment object with two assays "N" and "Q" respectively storing raw counts and quantitative values.
        Seqnames store path of the file the value.
}
\description{
Read EMOTE data from a ZIP archive containing multiple EmoteConv output directories
}
\author{
Julien Prados
}
