% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantif.R
\name{EMOTE_write_quantif}
\alias{EMOTE_write_quantif}
\title{Helper function to write an EMOTE quantification file}
\usage{
EMOTE_write_quantif(covPos, covNeg, R, file)
}
\description{
Helper function to write an EMOTE quantification file
}
