% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_detect.R
\name{p.combine}
\alias{p.combine}
\title{Combine P-values of several replicates with Fisher's method}
\usage{
p.combine(P)
}
\arguments{
\item{P}{a matrix of P value with as many column as replicate}
}
\value{
the numeric vector of combined p-values
}
\description{
Combine P-values of several replicates with Fisher's method
}
