% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_detect.R
\name{tss.model}
\alias{tss.model}
\title{Compute TSS p.value for a given position according to a beta-binomial model}
\usage{
tss.model(k.pos, k.neg, n.pos, n.neg, p.lb = 1e-30)
}
\arguments{
\item{k.pos}{number of observed positive event in the tested condition}

\item{k.neg}{number of observed positive event in the backgroud condition}

\item{n.pos}{total number of event in the tested condition}

\item{n.neg}{total number of event in the backgroud condition}

\item{p.lb}{lower bound on the p-value}
}
\value{
p-value of the enrichment test in pos condition as compared to background condition
}
\description{
Compute TSS p.value for a given position according to a beta-binomial model
}
